requirements.txt
beautifulsoup4==4.5.1
cssutils==1.0.1
Django==1.9.7
django-allauth==0.27.0
django-analytical==2.2.2
django-bootstrap3==8.2.3
django-ckeditor==5.1.1
django-crispy-forms==1.6.0
django-extensions==1.7.4
django-impersonate==1.0.1
django-import-export==0.5.1
django-inline-svg==0.1.0
django-inlinecss==0.1.2
django-liststyle==0.2b0
django-private-storage==1.2
django-settings-export==1.1.0
djangorestframework==3.4.7
easy-thumbnails==2.3
gunicorn==19.4.5
html5lib==0.9999999
numpy==1.13.0
Pillow==4.1.1
psycopg2==2.6.2
raven==5.32.0
requests==2.11.1
scipy==0.19.1
six==1.10.0
tensorflow==1.2.1
